+++ src/App.jsx
@@ -3,7 +3,8 @@ import { routes } from "./routes/routes";
 import "./styles/global.scss";
 
 function App() {
-  return <RouterProvider router={routes} />;
+  return <RouterProvider router={routes}  />; 
+  
 }
 
 export default App;
