+++ src/components/important/index.js
@@ -1,0 +1,1 @@
+export { Important } from './important'
