+++ src/styles/global.scss
@@ -1,30 +1,3 @@
 @import "./reset.scss";
 @import "./variables.scss";
 
-.important {
-    margin: 0 ;
-    padding: 20px;
-    font-family: Arial, sans-serif;
-    display: flex;
-    flex-direction: column;
-}
-
-h2 {
-  font-weight: 500;
-  padding-left: 110px;
-  margin-bottom: 40px;
-}
-
-.all-content {
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-}
-
-.first-img {
-  padding-bottom: 37px;
-}
-
-h3 {
-  padding-bottom: 24px;
-}
